refactor(layout-v2): move profile to rail-bottom avatar, tighten padding
- Reduce the rail's horizontal padding (px-2 -> px-1.5) so it feels less bulky.
- Remove the Profile tab from the top category list and relocate it to the
  bottom of the rail (below the settings icon) as the profile avatar, with the
  same click/panel navigation the tab had.
- Move the streak / cores / reputation cluster below the avatar.
- Drop the avatar + name + stats block from the expanded Home panel (the stats
  now live only on the rail; create-post action stays).
- Extract StreakPopover into its own file and delete the now-unused
  SidebarHeaderStats component.
